Correctly handle missing target at joyride start
We need to prevent setting `isRunning` during `start()` if there is
no mounted target, just like we do now in `toggleTooltip()`.  

This introduced a small problem, because there was no good way to detect
that a start was attempted from within componentWillUpdate.
The `isRunning` state does not change, nor does the index or any other
state for that matter.  For this reason, I’ve set the action to `start`
within the `start()` method, which seems reasonable.  Then, we can check
if we’re not actually running.  That can be because there are no steps,
or because the step had a target which isn’t mounted.  We don’t have a 
callback for when trying to start with no steps, but we could easily
add that now as well.

This also sets a default value for `startIndex` instead of performing
the check down in the `setState`.  They are equivalent, but setting it
directly as a default seems more explicit, and this way we can check
for a mounted target.