Allow spotlight styles to be updated from outside
When <Joyride /> is used as third party lib in host react applications, there are several instances where you might need a more finer control on how things are displayed. This change allows developers to manipulate spotlight style if they want according to their needs.

In our case, we are using <Joyride /> with react-pdf to create highlighted section in the pdf. So we need finer control over the spotlight CSS being displayed in our app.